EG pings gone

+++ cogs/Everglow.py
@@ -11,9 +11,6 @@ sadboi = 642497143801905190
 #.luminary bot-commands
 kbotcom = 764610881513324574
 
-#//people
-weakado = 259409277482041344
-
 class everglow(commands.Cog):
 
 
@@ -369,14 +366,9 @@ class everglow(commands.Cog):
                 await ctx.send(random.choice(self.bot.mia_gif))
                 await ctx.message.delete()
         elif arg == "yiren":
-            if ctx.guild.id == luminary:
-                if ctx.channel.id == kbotcom:
-                    await ctx.send(f'<@{weakado}>, <@!{ctx.author.id}> is talking about Yiren :orange_heart:')
-                    await ctx.send(random.choice(self.bot.yiren_gif))
-                    await ctx.message.delete()
-                else:
-                    await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
-                    await ctx.message.delete()
+            if ctx.guild.id == luminary and ctx.channel.id != kbotcom:
+                await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
+                await ctx.message.delete()
             else:
                 await ctx.send(f'<@!{ctx.author.id}> is talking about Yiren :orange_heart:')
                 await ctx.send(random.choice(self.bot.yiren_gif))
